+++ test/Braintree.Tests.Integration/PaymentMethodNonceIntegrationTest.cs
@@ -129,6 +129,7 @@ namespace Braintree.Tests.Integration
             Assert.AreEqual("test_eci", info.EciFlag);
             Assert.AreEqual("1.0.2", info.ThreeDSecureVersion);
             Assert.AreEqual("test_xid", info.Xid);
+            Assert.IsNotNull(info.ThreeDSecureAuthenticationId);
         }
 
         [Test]